feat: Update for Bevy 0.14.0

+++ examples/benchmark.rs
@@ -1,6 +1,6 @@
 use std::ops::Range;
 
-use bevy::{math::vec2, prelude::*, window::WindowResolution};
+use bevy::{math::uvec2, prelude::*, window::WindowResolution};
 
 use bevy_simple_tilemap::prelude::*;
 
@@ -108,7 +108,7 @@ fn setup(
 ) {
     // Load tilesheet texture and make a texture atlas from it
     let texture = asset_server.load("textures/tilesheet.png");
-    let atlas = TextureAtlasLayout::from_grid(vec2(16.0, 16.0), 4, 1, Some(vec2(1.0, 1.0)), None);
+    let atlas = TextureAtlasLayout::from_grid(uvec2(16, 16), 4, 1, Some(uvec2(1, 1)), None);
     let texture_atlas = texture_atlases.add(atlas);
 
     // Set up tilemap
