+++ examples/two_tilemaps.rs
@@ -1,5 +1,5 @@
 use bevy::{
-    math::{ivec2, vec2},
+    math::{ivec2, uvec2},
     prelude::*,
     window::WindowResolution,
 };
@@ -179,7 +179,7 @@ fn setup(
 ) {
     // Load tilesheet texture and make a texture atlas from it
     let texture = asset_server.load("textures/tilesheet.png");
-    let atlas = TextureAtlasLayout::from_grid(vec2(16.0, 16.0), 4, 1, Some(vec2(1.0, 1.0)), None);
+    let atlas = TextureAtlasLayout::from_grid(uvec2(16, 16), 4, 1, Some(uvec2(1, 1)), None);
     let texture_atlas = texture_atlases.add(atlas);
 
     let mut tiles = Vec::new();
